backend/requirements.txt
Django
django-cors-headers
djangorestframework
django-extensions
psycopg2-binary
alpaca-py
requests